+++ server/src/weaverbird/backends/pandas_executor/pipeline_executor.py
@@ -108,10 +108,6 @@ class PipelineExecutionFailure(PipelineFailure):
     """Raised when an error happens during the execution of the pipeline"""
 
     def __init__(self, step: PipelineStep, index: int, original_exception: Exception):
-        super(PipelineExecutionFailure, self).__init__(
-            step_name=step.name,
-            index=index,
-            original_exception=original_exception
-        )
+        super().__init__(step_name=step.name, index=index, original_exception=original_exception)
         self.step = step
         self.index = index