+++ server/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Evolution: add a global `PipelineFailure` exception to make easier to catch translation and execution errors from supported backends. Moreover, Pypika backends can also raise a `PipelineTranslationFailure` which specify the error, the step name and the step index. 
+
 ## [0.47.1] - 2024-09-06
 
 ### Fixed